requirements.txt
django==3.2.6
gunicorn==20.1.0
psycopg2==2.9.1
django-auth-adfs==1.9.0
wmi-client-wrapper-py3==2018.3
django-notifications-hq==1.6.0
Pillow
openpyxl==3.0.9

